support for >2GB trim commands

+++ minebd/src/main/java/io/minebox/nbd/ep/ExportProvider.java
@@ -16,7 +16,7 @@ public interface ExportProvider {
 
     void flush() throws IOException;
 
-    void trim(long offset, int length) throws IOException;
+    void trim(long offset, long length) throws IOException;
 
     default boolean supportsClientFlags(int clientFlags) {
         return true; //todo find out what those actually do
